# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VulnBank API",
    "description": "VulnBank — A banking API for demonstration purposes",
    "version": "1.0.0",
    "contact": {
      "name": "VulnBank Support",
      "email": "support@vulnbank.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:7777",
      "description": "Local development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "apiKey": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "username": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "fullName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "username": { "type": "string" },
                    "apiKey": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "409": { "description": "Username already exists" }
        }
      }
    },
    "/api/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "List all accounts",
        "operationId": "listAccounts",
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accounts": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Account" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Create account",
        "operationId": "createAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "integer" },
                  "type": { "type": "string", "enum": ["checking", "savings"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Account" }
              }
            }
          }
        }
      }
    },
...
    "/api/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "List transactions",
        "operationId": "listTransactions",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "ssn",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by SSN"
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["completed", "pending", "failed"] },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Transaction" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Transactions"],
        "summary": "Create transaction",
        "operationId": "createTransaction",
        "security": [{ "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fromAccountId", "toAccountId", "amount"],
                "properties": {
                  "fromAccountId": { "type": "integer" },
                  "toAccountId": { "type": "integer" },
                  "amount": { "type": "number" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get transaction by ID",
        "operationId": "getTransaction",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "404": { "description": "Transaction not found" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Transactions"],
        "summary": "Delete transaction",
        "operationId": "deleteTransaction",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Transaction deleted" },
          "404": { "description": "Transaction not found" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/balance": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get account balance",
        "operationId": "getBalance",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "schema": { "type": "string" },
            "description": "API key (alternative)"
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "integer" },
                    "accounts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "accountNumber": { "type": "string" },
                          "balance": { "type": "number" },
                          "type": { "type": "string" }
                        }
                      }
                    },
                    "totalBalance": { "type": "number" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/transfer": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Transfer funds",
        "operationId": "transferFunds",
        "security": [{ "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fromAccountId", "toAccountId", "amount"],
                "properties": {
                  "fromAccountId": { "type": "integer" },
                  "toAccountId": { "type": "integer" },
                  "amount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer successful"
          },
          "400": { "description": "Insufficient funds" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "operationId": "getCurrentUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "operationId": "updateUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "fullName": { "type": "string" },
                  "phone": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "operationId": "deleteUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/users/{id}/cards": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user credit cards",
        "operationId": "getUserCards",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "User credit cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cards": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CreditCard" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/users/search": {
      "post": {
        "tags": ["Users"],
        "summary": "Search users",
        "operationId": "searchUsers",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": { "type": "string" },
                    "sql": { "type": "string" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "username": { "type": "string" },
                          "email": { "type": "string" },
                          "fullName": { "type": "string" },
                          "matchedQuery": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/users/export": {
      "get": {
        "tags": ["Users"],
        "summary": "Export user data",
        "operationId": "exportUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": { "type": "string" },
            "description": "URL to fetch export data from"
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["json", "csv"] },
            "description": "Export format"
          }
        ],
        "responses": {
          "200": { "description": "Exported user data" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Account": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "accountNumber": { "type": "string" },
          "balance": { "type": "number" },
          "type": { "type": "string", "enum": ["checking", "savings"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "fromAccountId": { "type": "integer" },
          "toAccountId": { "type": "integer" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["completed", "pending", "failed"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string" },
          "fullName": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
...
    }
  }
}